requirements.txt
numpy==1.24.2
numpy==1.19.2
numpy==1.23.5
pandas==1.1.3
pandas==1.4.4
pandas==1.5.2
Pillow==8.0.1
Pillow==9.3.0
Pillow==9.5.0
scikit_learn==1.3.dev0
scikit_learn==0.23.2
scikit_learn==1.0.2
scikit_learn==1.1.3
streamlit==1.23.1
